Update examples/README.md
We need to update the `examples/README.md` to reflect recent ClimaComms changes to run on the Cluster. 

We don't use the env variable `CLIMACORE_DISTRIBUTED` anymore, but we need to specify `CLIMACOMMS_CONTEXT` instead. Even though ClimaComms should detect automatically if we're running on a distributed MPI env, I tried this via `sbatch` and if we don't specify this env variable, we run multiple single-process runs, rather than a distributed run.