Add initial support for script-based interface
This commit adds a script-based interface to ClimaAtmos (in alternative
to the driver.jl). Clarity, composability, and separation of concerns
are the key advantages of this interface. Note that this interface does
not make ClimaAtmos fully modular because of the way tendencies, cache,
and pre-computed values are handled. We are not aiming at that at the
moment.

This commit introduces enough infrastructure to support the simple
`single_column_hydrostatic_balance_ft64` case.

A test is added to verify that the output is the same. The test
currently uses an external HDF5 utility because comparing the integrator
objects in particularly tedious.

A good fraction of the new code is there to ensure
backward-compatibility.