Make ALL_DIAGNOSTICS an implementation detail
Currently, the diagnostic variables are stored in a dictionary. We might
want to be able to change the internal representation in the future. For
that, we provide an interface to the database of known diagnostic
variables through `set_diagnostic_variable!` and
`get_diagnostic_variable`. If we want to change the internal
representation in the future, we will be able to do so easily.