Clean up Github Actions/buildkite
This PR:

- ensures that Github action workflows are killed when a new commit is
pushed to the same branch/PR (following
https://stackoverflow.com/a/67939898)

- removes the need for the test environment from builkite jobs. The test
environment is used only for one job. When the depot is cold,
instantiating the test environment takes several minutes (when we have a
good depot it takes 1-2 minutes). Given that instantiating the
environment has to run before all the tests, this leads to unnecessary
latency.

- adds a new GitHub workflow 'lookahead'. The lookahead workflow is
non-required to merge PRs and runs tests and one simulation (the one
removed from the buildkite run) with julia 1.9 and julia 1.10 and the
latest dependencies.

- updates checkout actions to v4