Add AtmosSimulation
This commit introduces breaking changes in the interface.

* get_integrator is now replaced with get_simulation

The immediate reason for this change is that when we have diagnostics,
we also have a bunch of writers that open files and keep them open for
all the writing operations. We must ensure that the files are properly
closed at the end of the simulation. Hence, the integrator is no longer
enough. We could find a way to add this piece of information to the
cache, but now that the cache is more rigid, this becomes a pain.

In any case, we want to work on the atmos interface, especially for
coupler compatibility. At the moment, AtmosSimulation is a thin wrapper
that contains only the integrator at the writers. But this object will
extended to contain more useful and general information.

* simulation was removed from the cache

Now that simulation is its own object, it doesn't need to be in the
cache anymore