Disable compression for NetCDF files
By default, NetCDF files were compressed on the fly. This had no
measurable performance impact when the file was synced immediately (or
maybe compression happened at sync?). Now that the sync is decoupled
from output, the io benchmark seems to be much closer to the theoretical
limit set by NCDatasets