Sync NetCDF file at every write
This change essentially disables the internal buffer in the NetCDF
library. With this change, every time something has to be written to
disk, it is immediately written (as opposed to buffering writes).

This ensures that datasets are always written and no data is lost, but
comes to a performance cost.

This was already the default behavior for GPU runs, so this change only
affects CPU runs