Fix compat and add downgrade action
The compats in ClimaAtmos were largely wrong. In this commit, I fix them. I also
add a new action that checks that our tests pass with the oldest possible
combination of packages that is advertised to be compatible with ClimaAtmos.

Since we required ClimaComms 0.6, I also removed all the checks on that version
for the import_required_backends