Add test for dependencies
Adding new dependencies to a Julia project has downsides. First, every
dependency increases the precompilation and import time of a package. For
complex packages, such as CUDA, this translates in several minutes of
precompilation time that every single user of ClimaAtmos (including our CI
runners) always pays. Second, dependencies can increase the cost of maintaining
ClimaAtmos and the surface area of things that can go wrong (a buggy update in a
dependency can break ClimaAtmos). Third, dependencies increase the complexity of
the project's environment, potentially causing compatibility issues and
restricting the versions we can use.

This new test follows the spirit of the regression tests and tries to capture
the cost of adding a new dependency by having developers explicitly come here
and