Remove t_start
The diagnostic module was using `t_start`, but this was under the wrong
assumption.

I was under the impression that integrator.t restarts from 0 upon
restart, so that we needed to manually offset the time.

I think I got this idea from ClimaLand, where t_start is used often.

Using t_start was (probably) leading to incorrect dates being written in
NetCDF files when the simulation was restarted.