Automatically detect restart file
This commit adds a new option, `detect_restart_file`. When
`detect_restart_file` is true, ClimaAtmos looks at the previous
`output_XXXX` output folder and tries to see if there's a restart file
in it. If yes, use it. This behavior is overridden when a restart file
is explicitely passed.

This commit uses and re-implments the same logic implemented in
ClimaUtilities.OutputPathGenerator. We might want to move that logic to
ClimaUtilities at some point.