Add support for saving checkpoints by calendar dates
This commit adds support for providing a calendar frequency to the function that
saves to the state to disk. This is required to ensure that diagnostics and
states are synced and output at the same time. To enable this, I use the same
infrastructure used by diagnostic module. Everything was already there, I just
needed to add the YAML parsing.