Do not add ᶜ∇²tke⁰ to cache if not needed
For the diagnostic EDMF case, `ᶜ∇²tke⁰` is only used when
`prognostic_tke` is true. This commit ensures that the variable is added
to the cache only in that case. Prior to this commit, `ᶜ∇²tke⁰` was
added unconditionally and would never be updated, so that it was always
full with random numbers (being initialized with similar).