Check that AtmosModel is the same across restarts
This commit adds the hash of the AtmosModel instance used to prepare a
simulation in the checkpoint HDF5 files. This is then used to check that
the AtmosModel is the "same" (up to a hash) across restarts.

A test is added to check that this mechanism works. In the test, I
produced a checkpoint and read it with the same configuration with one
value changed (insolation).