Update insolation_tuple for boxes too
While `insolation_tuple` seems to be currently used only in one
function, it is stored in the cache and available everywhere. For
spheres, `insolation_tuple` was updated in the radiation callback, but
that was not the case for other configurations. This commit ensures that
`insolation_tuple` is always updated. For non-spheres, we assume
insolation at Greenwhich (lat = long = 0).

This commit also unifies a little bit the logic between the two cases
(with and without sphere).