Add periods_are_commensurate
`periods_are_commensurate` is a function to check whether two
`Dates.Periods` are compatible, meaning one is an integer multiple of
the other. This is a little tricky to accomplish because some periods
don't have a fixed length: for example, one Month can be 28, 29, 30, and
31 days.

In `periods_are_commensurate`, I handle all the cases that can be
handled directly (e.g., periods with fixed length), and hardcoded some
special, but relevant cases. The most important of these cases is when
we have a Month.

Not all the cases are covered, but we can add them when/if they become
relavant.