Check that dt_rad and dt_checkpoint are compatible
If we want our restarts to be reproducible, we have to worry about
callbacks. Callbacks contain information about when they were last
called, information that is not saved to checkpoint. This is
particularly relevant for radiation. For example, if we save checkpoints
every day, but call radiation every 15 h, we will have that restarting a
simulation leads to a different evolution compared to not doing that.
This is because radiation is called at different times:

Non-restarted simulation
t = 15h: radiation
t = 24h: checkpoint
t = 30h: radiation

Restarted simulation:
t = 15h: radiation
t = 24h: checkpoint
t = 39h: radiation

This commit adds a warning that a given simulation will not be
reproducible when radiation and checkpoints frequencies are not
commensurate.