Add check for accumulated diagnostics and restarts
This commit adds a warning when we can detect checkpointing and
diagnostics being out of sync. ClimaAtmos now detects what frequencies
are involved and print them out, e.g.
```
[ Info: Saving accumulated diagnostics to disk with frequency: 2 days, 1 day
```
Then, it checks that all these frequencies evenly divide the
checkpointing frequency. If this does not happen, a warning is produced.