Make the default config table searchable in docs
The content of the old config.md was moved into config_no_table.md
with the include("config_table.jl") at the bottom removed.

Now, config_table.jl, creates the config.md by reading
config_no_table.md and copying the lines into it. Then,
it generates the pretty table and adds it to the end of the file.

This allows the table to be searchable, but it does not require users
to manually update the docs when the default config yaml is modified.

Fix formatting

Remove generated config.md from git index