Remove start_date from cache
`start_date` is constant and only used to initialize diagnostics and in
the computation of insolation. Here, I remove it from the cache. For
insolation, I moved it to the constructor. In the future, once we have
the integer time, we will be able to remove it from there as well.