Remove WTE from cache
This commit removes the `WallTimeEstimate` from the cache and moves it
to an isolated place. In the process, I refactored the struct to split
reporting with updating, so that reporting can be done with any
frequency/schedule desired using the same Schedule infrastructure used
by other functions/diagnostics as well.