Move aerosols to its own type
Aerosols were handles by parsing directly the YAML arguments. This led
to implicit assumptions, and required aerosols to be handled
differently. For example, we had to pass the list `aerosol_names` to
`build_cache` even when no aerosol was required.

This commit moves aerosol to its own type so that we can follow a more
composable approach. It is still not fully possible to swap in a custom,
user-defined type because the portion of code in RRTGMPInterface makes
assumption on the names of the aerosols. I add error checking for that