Consolidate Manifests
Manifests.toml are a burden to maintain and add a lot of git diff noise.
The one benefit of Manifests.toml is that they help with tracking
version dependencies.

However:
- For `docs`, the precise version of dependencies is not important,
there is pretty much no version-dependent code being executed there
- `perf` and `examples` can be consolidated into a single `.buildkite`
environment. This environment only has to have one manifest checked, the
one for which buildkite is run (and not one for each version of Julia).
- The `perf` environment can be removed all together. It is a more niche
environment: if reproducing a specific result is required, developers
can use the `.buildkite` Manifest. If not, developers can also consider
keeping the relevant tools (mainly JET) in the base environment.

This is the same pattern we have in other repos such as ClimaCore.