Remove dt_save_to_sol
`dt_save_to_sol` copies the state to `integrator.sol`. For long
simulations, this can be an expensive and unnecessary operation.

This used to be more useful in the past (before diagnostics were a
thing), but now it is only used for conservation checks.

This commit removes the option from the configurable options.

This is a breaking change.