Fix blank mse_summary
The mse_summary.jl script always prints no summaries or error.
This is because the reproducibility files are now placed in a
seperate subdir. This commit updates the `get_computed_mses` default
value for `subfolder` to reflect this

Second fix

Add a filter for the prog_state, as it is the last in the dir,
and t`get_computed_mse` assumes the last file is the
latest comparison,.

Add tests